RI-7630: move bootstrap file into proper folder
