+++ redisinsight/ui/src/styles/custom/dark_theme/index.ts
